+++ code/graphics/uniforms.h
@@ -18,5 +18,10 @@ void convert_model_material(model_uniform_data* data_out,
 							const vec3d& scale,
 							size_t transform_buffer_offset);
 
+void convert_shadow_material(shadow_uniform_data* data_out,
+							 const matrix4& model_transform,
+							 const vec3d& scale,
+							 size_t transform_buffer_offset);
+
 }
 }
